making some updates to personality class and json
