+++ scripts/cat/cats.py
@@ -25,6 +25,7 @@ from scripts.cat_relations.relationship import Relationship
 from scripts.game_structure import image_cache
 from scripts.event_class import Single_Event
 from .thoughts import Thoughts
+from .personality import Personality
 from scripts.cat_relations.inheritance import Inheritance
 
 
@@ -2945,140 +2946,7 @@ class Cat():
 #                               PERSONALITY CLASS                              #
 # ---------------------------------------------------------------------------- #
 
-class Personality():
-    """Hold personality information for a cat, and functions to deal with it """
-    facet_types = ["lawfulness", "sociability", "aggression", "stability"]
-    facet_range = [0, 16]
-    
-    with open("resources/dicts/traits/trait_ranges.json", "r") as read_file:
-        trait_ranges = ujson.loads(read_file.read())
-    
-    def __init__(self, trait:str=None, kit_trait:bool=False, lawful:int=None, social:int=None, aggress:int=None, 
-                 stable:int=None) -> Personality:
-        """If trait is given, it will randomize facets within the range of the trait. It will ignore any facets given. 
-            If facets are given and no trait, it will find a trait that matches the facets. NOTE: you can give
-            only some facets: It will randomize any you don't specify.
-            If both facets and trait are given, it will use the trait if it matched the facets. Otherwise it will
-            find a new trait."""
-        self._law = 0
-        self._social = 0
-        self._aggress = 0
-        self._stable = 0
-        self.trait = None
-        self.kit = kit_trait #If true, use kit trait. If False, use normal traits. 
-        
-        if self.kit:
-            trait_type_dict = Personality.trait_ranges["kit_traits"]
-        else:
-            trait_type_dict = Personality.trait_ranges["normal_traits"]
-        
-        _tr = None
-        if trait and trait in trait_type_dict:
-            # Trait-given init
-            self.trait = trait
-            _tr = trait_type_dict[self.trait]
-        
-        # Set Facet Values
-        # The priority of is: 
-        # (1) Given value, from parameter. 
-        # (2) If a trait range is assigned, pick from trait range
-        # (3) Totally random. 
-        if lawful is not None:
-            self._law = Personality.adjust_to_range(lawful)
-        elif _tr:
-            self._law = randint(_tr["lawfulness"][0], _tr["lawfulness"][1])
-        else:
-            self._law = randint(Personality.facet_range[0], Personality.facet_range[1])
-            
-        if social is not None:
-            self._social = Personality.adjust_to_range(social)
-        elif _tr:
-            self._social = randint(_tr["sociability"][0], _tr["sociability"][1])
-        else:
-            self._social = randint(Personality.facet_range[0], Personality.facet_range[1])
-            
-        if aggress is not None:
-            self._aggress = Personality.adjust_to_range(aggress)
-        elif _tr:
-            self._aggress = randint(_tr["aggression"][0], _tr["aggression"][1])
-        else:
-            self._aggress = randint(Personality.facet_range[0], Personality.facet_range[1])
-            
-        if stable is not None:
-            self._stable = Personality.adjust_to_range(stable)
-        elif _tr:
-            self._stable = randint(_tr["stability"][0], _tr["stability"][1])
-        else:
-            self._stable = randint(Personality.facet_range[0], Personality.facet_range[1])
-                
-        # If trait is still empty, or if the trait is not valid with the facets, change it. 
-        if not self.trait or not self.is_trait_valid():
-            self.choose_trait()
-                
-    def __repr__(self) -> str:
-        """For debugging"""
-        return f"{self.trait}: lawfulness {self.lawfulness}, aggression {self.aggression}, sociability {self.sociability}, stablity {self.stability}"
-    
-    def get_facet_string(self):
-        """For saving the facets to file."""
-        return f"{self.lawfulness},{self.sociability},{self.aggression},{self.stability}"
-    
-    def __getitem__(self, key):
-        """Alongside __setitem__, Allows you to treat this like a dictionary if you want. """
-        return getattr(self, key)
-    
-    def __setitem__(self, key, newval):
-        """Alongside __getitem__, Allows you to treat this like a dictionary if you want. """
-        setattr(self, key, newval)
 
-    # ---------------------------------------------------------------------------- #
-    #                               PROPERTIES                                     #
-    # ---------------------------------------------------------------------------- #
-    
-    @property
-    def lawfulness(self):
-        return self._law
-    
-    @lawfulness.setter
-    def lawfulness(self, new_val):
-        """Do not use property in init"""
-        self._law = Personality.adjust_to_range(new_val)
-        if not self.is_trait_valid():
-            self.choose_trait()
-            
-    @property
-    def sociability(self):
-        return self._social
-    
-    @sociability.setter
-    def sociability(self, new_val):
-        """Do not use property in init"""
-        self._social = Personality.adjust_to_range(new_val)
-        if not self.is_trait_valid():
-            self.choose_trait()
-            
-    @property
-    def aggression(self):
-        return self._aggress
-    
-    @aggression.setter
-    def aggression(self, new_val):
-        """Do not use property in init"""
-        self._aggress = Personality.adjust_to_range(new_val)
-        if not self.is_trait_valid():
-            self.choose_trait()
-            
-    @property
-    def stability(self):
-        return self._stable
-    
-    @stability.setter
-    def stability(self, new_val):
-        """Do not use property in init"""
-        self._stable = Personality.adjust_to_range(new_val)
-        if not self.is_trait_valid():
-            self.choose_trait()
-            
 
     # ---------------------------------------------------------------------------- #
     #                               METHODS                                        #
@@ -3183,6 +3051,8 @@ class Personality():
         else:
             #This will only trigger if they have the same personality. 
             return None
+
+## end personality class
 
 # Twelve example cats
 def create_example_cats():
